normalization based on the global distribution
For features with IQR=0, the interquantile_normalization falls, then the minmax_normalization is executed. The output file is in both formats: .csv and .h5.

Also includes the imputing for missing values through either: by median values or by zero values. To choose one imputing method is enough with editing the nan_cols dictionary.